Fix missing multi-editing support in MinMaxRangeDrawer
The MinMaxRangeDrawer was not correctly supporting multi-object-editing. When editing multiple objects, they would share the same values after selecting them. With this change, multi-editing is supported. The code has been refactored and slightly polished as well. When moving the clamped slider, float fields are now unfocused so that they correctly update while moving the slider.